Functions for calculating bonus points are placed in func_bonus.py
